Extract event push helpers to break circular import
